# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/domain": {
            "get": {
                "description": "get all domains",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "get all domain",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.Domain"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create new domain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "domain create",
                "parameters": [
                    {
                        "description": "necessary item for create new",
                        "name": "DomainCreateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDomainRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/domain/:id": {
            "delete": {
                "description": "delete domain with id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "domain"
                ],
                "summary": "domain delete",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/zone": {
            "get": {
                "description": "get all zone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "zone"
                ],
                "summary": "get all zones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CacheZone"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create new zone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "zone"
                ],
                "summary": "create zone",
                "parameters": [
                    {
                        "description": "necessary item for create new",
                        "name": "CreateCacheZone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddCacheZoneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AddCacheZoneRequest": {
            "type": "object",
            "required": [
                "inactiveSecond",
                "maxSizeMB",
                "path",
                "zoneName"
            ],
            "properties": {
                "inactiveSecond": {
                    "type": "integer"
                },
                "maxSizeMB": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "zoneName": {
                    "type": "string"
                }
            }
        },
        "dto.CreateDomainRequest": {
            "type": "object",
            "required": [
                "address",
                "name",
                "zoneId"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "cacheKey": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rateLimit": {
                    "$ref": "#/definitions/dto.rateLimit"
                },
                "zoneId": {
                    "type": "integer"
                }
            }
        },
        "dto.Domain": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.rateLimit": {
            "type": "object",
            "properties": {
                "burst": {
                    "type": "integer"
                },
                "maxSize": {
                    "type": "string"
                },
                "rate": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "models.CacheZone": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inactive": {
                    "type": "string"
                },
                "maxSize": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        }
    }
}